fix cross compilation
